feat: move req-analysis page
